added manager sign up page
