+++ dms/dealership-site/src/Menu.css
@@ -1,14 +1,14 @@
 a:link {
-  text-decoration: none;
+	text-decoration: none;
 	color: white;
 }
 
 .LogoBar {
-  background-color: #0D1117;
-  width: 100vw;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
+	background-color: #0D1117;
+	width: 100vw;
+	display: flex;
+	justify-content: space-between;
+	align-items: flex-end;
 }
 
 .AcctBtns {
@@ -19,56 +19,80 @@ a:link {
 }
 
 .Button {
-  color: white;
-  font-size: x-large;
-  height: 40px;
-  width: 9vw;
-  display: flex;
-  justify-content: center;
-  align-items: center;
+	color: white;
+	font-size: x-large;
+	height: 40px;
+	width: 9vw;
+	display: flex;
+	justify-content: center;
+	align-items: center;
 }
 
 .Login {
-  background-color: #2062c6;
+	background-color: #2062c6;
 }
 
 .Login:hover {
-  background-color: #417cd5;
+	background-color: #417cd5;
 }
 
-.SignUp {
-  background-color: #8d39e1;
+.DropDownLinkHidden {
+	visibility: hidden;
+	opacity: 0;
+	display: none;
+}
+
+.DropDownLink {
+	background-color: #8d39e1;
+	color: white;
+	font-size: large;
+}
+
+.SignUpBtn {
+	z-index: 1;
+	background-color: #8d39e1;
+	color: white;
+	font-size: x-large;
+	width: 9vw;
+	display: flex;
+	flex-direction: column;
+	justify-content: flex-start;
+	align-items: center;
+}
+
+.Collapsed {
+	height: 40px;
 }
 
 .NavBarAndContent {
-  display: flex;
+	display: flex;
 	flex-direction: column;
 }
 
 .NavBar {
-  background-color: #223148;
-  color: white;
-  width: 100vw;
-  min-width: 120px;
-  height: 5vh;
+	background-color: #223148;
+	color: white;
+	width: 100vw;
+	min-width: 120px;
+	height: 5vh;
 }
 
 .NavBtn {
-  background-color: #223148;
-  color: white;
-  text-wrap: nowrap;
-  border: none;
-  width: 15vw;
-  min-width: 120px;
-  height: 5vh;
-  min-height: 25px;
-  font-size: large;
+	background-color: #223148;
+	color: white;
+	text-wrap: nowrap;
+	border: none;
+	width: 15vw;
+	min-width: 120px;
+	height: 5vh;
+	min-height: 25px;
+	font-size: large;
 }
 
 .NavBtn:hover {
-  background-color: #324869;
+	background-color: #324869;
 }
 
 .Hidden {
-  display: none;
+	display: none;
 }
